Attack Range indicator
Created script to show the attack range of a unit when you control click it
Added a coroutine to the inital loading function so it better shows how the game is loading.
Changed catapult to a mage, as i dont have any art for a catapult yet. (still functions about teh same)
Added sprited and animations for teh archer and mage(still need to add something for when they shoot)